requirements.txt
roboflow
diffusers 
transformers
peft
wandb 
omegaconf
ml-collections
controlnet-aux
pytorch-msssim
optuna
torchmetrics
torchmetrics[image]